% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{add_daily_count_frum_cumulative}
\alias{add_daily_count_frum_cumulative}
\title{add_daily_count create a daily number of counts from a cumulative count  -   step 5 /5  in the "create daile counts from cumulative counts" chain}
\usage{
add_daily_count_frum_cumulative(df, date, cumulative, daily, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_daily_count create a daily number of counts from a cumulative count  -   step 5 /5  in the "create daile counts from cumulative counts" chain
}
\examples{

data <- tribble(
  ~"date_report", ~"age", ~"sex", ~"cumulatif1", ~"cumulatif2",
  "2021-01-01", 18, "M", 10, 1,
  "2021-01-02", 18, "M", 9, 2,
  "2021-01-04", 18, "M", 12, 4,
  "2021-01-01", 19, "M", 10, 1,
  "2021-01-02", 19, "M", 9, 2,
  "2021-01-04", 19, "M", 12,4,
  "2021-01-01", 19, "F", 10,1,
  "2021-01-02", 19, "F", 9,2,
  "2021-01-04", 19, "F", 12, 4
) \%>\%
  mutate(date_report=as.Date(date_report))
data \%>\%
  fix_cumulative(., date_report, cumulatif1, age, sex) \%>\%   # fix first cumulatif pour assurer qu'oon a pas de baisse de cmulatif
  fix_cumulative(., date_report, cumulatif2, age, sex)  \%>\% # fix second cumulatif pour assurer qu'oon a pas de baisse de cmulatif
  add_slope(., date_report, cumulatif1, age, sex) \%>\%  # add slope for cumulatif1 for fill in missing dates
  add_slope(., date_report, cumulatif2, age, sex)  \%>\% # add slope for cumulatif2 for fill in missing dates
  add_missing_dates(., date_report, age, sex) \%>\%
  fill_cumulative_using_slope(., date_report, cumulatif1, age, sex) \%>\%
  fill_cumulative_using_slope(., date_report, cumulatif2, age, sex) \%>\%
  add_daily_count_frum_cumulative(., date_report, cumulatif1, daily1, age,sex) \%>\%
  add_daily_count_frum_cumulative(., date_report, cumulatif2, daily2, age,sex)

}
