% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{add_missing_dates}
\alias{add_missing_dates}
\title{add missing dates add missing rows   -   step 3 in the "create daile counts from cumulative counts" chain}
\usage{
add_missing_dates(df, date, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add missing dates add missing rows   -   step 3 in the "create daile counts from cumulative counts" chain
}
