% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{add_slope}
\alias{add_slope}
\title{add_slope evaluates how many counts were added daily between 2 cumulative values -   step 2 in the "create daile counts from cumulative counts" chain}
\usage{
add_slope(df, date, cumulative, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_slope evaluates how many counts were added daily between 2 cumulative values -   step 2 in the "create daile counts from cumulative counts" chain
}
