% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cronjob_data.R
\name{create_date_report_from_datetimes}
\alias{create_date_report_from_datetimes}
\title{Title}
\usage{
create_date_report_from_datetimes(datetimes, csvs)
}
\arguments{
\item{csvs}{}
}
\value{

}
\description{
Title
}
