% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{fill_cumulative_using_slope}
\alias{fill_cumulative_using_slope}
\title{fill_cumulative_using slope replaces cumulative values from previous missing days using the previously calculated slope  -   step 4 in the "create daile counts from cumulative counts" chain}
\usage{
fill_cumulative_using_slope(df, date, cumulative, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
fill_cumulative_using slope replaces cumulative values from previous missing days using the previously calculated slope  -   step 4 in the "create daile counts from cumulative counts" chain
}
