% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_data.R
\name{fill_missing_dates_and_create_daily_counts_for_mtl_data}
\alias{fill_missing_dates_and_create_daily_counts_for_mtl_data}
\title{Title}
\usage{
fill_missing_dates_and_create_daily_counts_for_mtl_data(mtl_data)
}
\arguments{
\item{rls_data}{}
}
\value{

}
\description{
Title
}
