% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rls_data.R
\name{fill_missing_dates_and_create_daily_counts_for_rls_data}
\alias{fill_missing_dates_and_create_daily_counts_for_rls_data}
\title{fill_missing_dates_and_create_daily_counts_for_rls_data}
\usage{
fill_missing_dates_and_create_daily_counts_for_rls_data(rls_data)
}
\arguments{
\item{rls_data}{}
}
\value{

}
\description{
fill_missing_dates_and_create_daily_counts_for_rls_data
}
