% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{fix_cumulative}
\alias{fix_cumulative}
\title{fix_cumulative ensures than no cumulative count is lower than a previous value.  step 1 in the "create daile counts from cumulative counts" chain}
\usage{
fix_cumulative(df, date, cumulative, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
fix_cumulative ensures than no cumulative count is lower than a previous value.  step 1 in the "create daile counts from cumulative counts" chain
}
