% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_date.R
\docType{data}
\name{format_anglais}
\alias{format_anglais}
\title{format de date en anglais}
\format{
An object of class \code{character} of length 1.
}
\usage{
format_anglais
}
\description{
format de date en anglais
}
\keyword{datasets}
