% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_date.R
\name{format_chiffre}
\alias{format_chiffre}
\title{convert number to character, with a " " bigmark separator}
\usage{
format_chiffre(chiffre, big.mark = " ", ...)
}
\description{
convert number to character, with a " " bigmark separator
}
