% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_date.R
\docType{data}
\name{format_date_francais}
\alias{format_date_francais}
\title{format date as string 10 novembre 2021}
\format{
An object of class \code{character} of length 1.
}
\usage{
format_date_francais
}
\description{
format date as string 10 novembre 2021
}
\keyword{datasets}
