% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_date.R
\name{format_piasse}
\alias{format_piasse}
\title{convert number to character with a dollar sign}
\usage{
format_piasse(chiffre, big.mark = " ", prefix = "", suffix = " $", ...)
}
\description{
convert number to character with a dollar sign
}
