% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_date.R
\name{format_pourcent}
\alias{format_pourcent}
\title{convert number to character with a percentage sign}
\usage{
format_pourcent(chiffre, accuracy = 0.1, suffix = "\%", ...)
}
\description{
convert number to character with a percentage sign
}
