% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_data.R
\name{get_current_municipal}
\alias{get_current_municipal}
\title{get_current_tableau_rls_new_from_inspq}
\usage{
get_current_municipal()
}
\value{

}
\description{
get_current_tableau_rls_new_from_inspq
}
\examples{



}
