% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rls_data.R
\name{get_current_tableau_rls_new_from_inspq}
\alias{get_current_tableau_rls_new_from_inspq}
\title{get_current_tableau_rls_new_from_inspq}
\usage{
get_current_tableau_rls_new_from_inspq()
}
\value{

}
\description{
get_current_tableau_rls_new_from_inspq
}
