% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_data.R
\name{get_historical_municipal_from_github}
\alias{get_historical_municipal_from_github}
\title{get_historical_municipal_from_github}
\usage{
get_historical_municipal_from_github()
}
\value{

}
\description{
get_historical_municipal_from_github
}
