% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inspq_covid19_hist.R
\name{get_inspq_covid19_hist}
\alias{get_inspq_covid19_hist}
\title{get_inspq_covid19_hist}
\usage{
get_inspq_covid19_hist()
}
\value{

}
\description{
get_inspq_covid19_hist
}
\examples{
get_inspq_covid19_hist()
}
