% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inspq_manual_data_hospits.R
\name{get_inspq_manual_data_hospits}
\alias{get_inspq_manual_data_hospits}
\title{Importer le dernier fichier manual_data de inspq et le nettoyer}
\usage{
get_inspq_manual_data_hospits()
}
\value{
data frame
}
\description{
Importer le dernier fichier manual_data de inspq et le nettoyer
}
\examples{
get_inspq_manual_data_hospits()
}
