% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inspq_manual_data_hospits.R
\name{get_inspq_manual_data_tableau_accueil}
\alias{get_inspq_manual_data_tableau_accueil}
\title{get_inspq_manual_data_tableau_accueil}
\usage{
get_inspq_manual_data_tableau_accueil()
}
\value{

}
\description{
get_inspq_manual_data_tableau_accueil
}
\examples{
get_inspq_manual_data_tableau_accueil()
}
