% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_data.R
\name{get_jeanpaulrsoucy_municipal}
\alias{get_jeanpaulrsoucy_municipal}
\title{get_jeanpaulrsoucy_municipal}
\usage{
get_jeanpaulrsoucy_municipal()
}
\value{

}
\description{
get_jeanpaulrsoucy_municipal
}
