% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rls_data.R
\name{get_jeanpaulrsoucy_tableau_rls_new}
\alias{get_jeanpaulrsoucy_tableau_rls_new}
\title{get_jeanpaulrsoucy_tableau_rls_new}
\usage{
get_jeanpaulrsoucy_tableau_rls_new()
}
\value{

}
\description{
get_jeanpaulrsoucy_tableau_rls_new
}
