% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_data.R
\name{get_raw_mtl_data}
\alias{get_raw_mtl_data}
\title{get_raw_rls_data - combine RLS data from 3 sources to get the best possible historical data}
\usage{
get_raw_mtl_data()
}
\value{

}
\description{
get_raw_rls_data - combine RLS data from 3 sources to get the best possible historical data
}
