% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rls_data.R
\name{get_raw_rls_data}
\alias{get_raw_rls_data}
\title{}
\usage{
get_raw_rls_data()
}
\value{

}
\description{
) %>%
    select(RSS, RLS, cumulative_cases, date_report) %>%
    filter(!is.na(cumulative_cases))
}
\details{

}
