% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccins_data.R
\name{get_raw_vaccination_data}
\alias{get_raw_vaccination_data}
\title{get_raw_vaccination_data downloads vaccination data from inspq}
\usage{
get_raw_vaccination_data()
}
\value{

}
\description{
get_raw_vaccination_data downloads vaccination data from inspq
}
