% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_data.R
\name{get_raw_variant_data}
\alias{get_raw_variant_data}
\title{get_raw_variant_data downloads variant data from JP's s3 bucket}
\usage{
get_raw_variant_data()
}
\value{

}
\description{
get_raw_variant_data downloads variant data from JP's s3 bucket
}
