% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_quebec_cas_par_css.R
\name{graph_css_bars}
\alias{graph_css_bars}
\title{Title}
\usage{
graph_css_bars(css_last_week = NULL)
}
\arguments{
\item{css_last_week}{OPTIONAL: if you provide a dataframe created by css_last_week then you don't need to redownload everything again.}
}
\value{

}
\description{
Title
}
