% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_hospit_deces_tests.R
\name{graph_deces_hospit_tests}
\alias{graph_deces_hospit_tests}
\title{graph_deces_hospit_tests}
\usage{
graph_deces_hospit_tests()
}
\value{

}
\description{
graph_deces_hospit_tests
}
\examples{
graph_deces_hospit_tests()
}
