% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_quebec_cas_par_age.R
\name{graph_quebec_cas_par_age_heatmap}
\alias{graph_quebec_cas_par_age_heatmap}
\title{graph_quebec_cas_par_age_heatmap}
\usage{
graph_quebec_cas_par_age_heatmap()
}
\value{

}
\description{
graph_quebec_cas_par_age_heatmap
}
