% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_quebec_cas_pour_mtl.R
\name{graph_quebec_cas_par_mtl_heatmap}
\alias{graph_quebec_cas_par_mtl_heatmap}
\title{graph_quebec_cas_par_mtl_heatmap}
\usage{
graph_quebec_cas_par_mtl_heatmap(mtl_data = NULL)
}
\value{

}
\description{
graph_quebec_cas_par_mtl_heatmap
}
