% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_quebec_cas_par_region.R
\name{graph_quebec_cas_par_region}
\alias{graph_quebec_cas_par_region}
\title{graph quebec cas pasr région}
\usage{
graph_quebec_cas_par_region()
}
\value{

}
\description{
graph quebec cas pasr région
}
