% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_inspq_covid19_hist.R
\name{load_inspq_covid19_hist}
\alias{load_inspq_covid19_hist}
\title{load_inspq_covid19_hist}
\usage{
load_inspq_covid19_hist()
}
\value{

}
\description{
load_inspq_covid19_hist
}
\examples{
load_inspq_covid19_hist()
}
