% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_inspq_manual_data.R
\name{load_inspq_manual_data}
\alias{load_inspq_manual_data}
\title{Importer le dernier fichier manual_data de inspq et le nettoyer}
\usage{
load_inspq_manual_data(path = "~/cronjob/inspq-manual-data/")
}
\arguments{
\item{path}{string}
}
\value{
data frame
}
\description{
Importer le dernier fichier manual_data de inspq et le nettoyer
}
