% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mes4couleurs}
\alias{mes4couleurs}
\title{mes4 couleurs de paliers d'alertes}
\format{
An object of class \code{character} of length 4.
}
\usage{
mes4couleurs
}
\description{
mes4 couleurs de paliers d'alertes
}
\keyword{datasets}
