% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggsave.R
\name{myggsave}
\alias{myggsave}
\title{myggsave sauvegarde les ggplot avec une hauteur et largeur par défaut..}
\usage{
myggsave(
  filename = default_name(plot),
  plot = last_plot(),
  height = 9,
  width = 12,
  dpi = 250,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
myggsave sauvegarde les ggplot avec une hauteur et largeur par défaut..
}
