% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{populations}
\alias{populations}
\title{population des health regions selon ishaberry repository}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 93 rows and 5 columns.
}
\usage{
populations
}
\description{
population des health regions selon ishaberry repository
}
\keyword{datasets}
