% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{populations_age}
\alias{populations_age}
\title{population des groupe d'age du Québec}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 2 columns.
}
\usage{
populations_age
}
\description{
population des groupe d'age du Québec
}
\keyword{datasets}
