% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{populations_age_montreal}
\alias{populations_age_montreal}
\title{population des groupe d'age de Montréal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
populations_age_montreal
}
\description{
population des groupe d'age de Montréal
}
\keyword{datasets}
