% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data}
\usage{
prep_data(data, group, type)
}
\arguments{
\item{data}{data frame}

\item{group}{nom de variable groupe, ex: health_region}

\item{type}{type de ce qui est compté, ex: cases ou deaths}
}
\value{

}
\description{
prep_data
}
\examples{
prep_data(cases_prov, province, type = cases )
}
