% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cronjob_data.R
\name{prepare_cronjob_data_nobindrows}
\alias{prepare_cronjob_data_nobindrows}
\title{prepare_cronjob_data_nobindrows}
\usage{
prepare_cronjob_data_nobindrows(
  mypath = "~/cronjob/tableau-rls/",
  skip = 0,
  n_max = Inf,
  locale = default_locale(),
  csv_date_start = -19,
  csv_date_end = -12,
  delim = ","
)
}
\arguments{
\item{delim}{}
}
\value{

}
\description{
prepare_cronjob_data_nobindrows
}
