% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert.R
\docType{data}
\name{quebec_lambert}
\alias{quebec_lambert}
\title{String to define the Quebec lambert projection}
\format{
An object of class \code{character} of length 1.
}
\usage{
quebec_lambert
}
\description{
Projection found at https://spatialreference.org/ref/epsg/nad83-quebec-lambert/
with the actual string found at # https://spatialreference.org/ref/epsg/nad83-quebec-lambert/proj4/
}
\keyword{datasets}
