% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{rls_population}
\alias{rls_population}
\title{population des RLS du Québec}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 2 columns.
}
\usage{
rls_population
}
\description{
population des RLS du Québec
}
\keyword{datasets}
