% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert.R
\docType{data}
\name{statistics_canada_lambert}
\alias{statistics_canada_lambert}
\title{String to define the Statistics Canada lambert projection
#' Projection found at https://spatialreference.org/ref/epsg/3347/
actual string from: https://spatialreference.org/ref/epsg/3347/proj4/
appears to fit the description on statistics canada here  https://www150.statcan.gc.ca/n1/pub/92-195-x/2011001/other-autre/mapproj-projcarte/m-c-eng.htm}
\format{
An object of class \code{character} of length 1.
}
\usage{
statistics_canada_lambert
}
\description{
String to define the Statistics Canada lambert projection
#' Projection found at https://spatialreference.org/ref/epsg/3347/
actual string from: https://spatialreference.org/ref/epsg/3347/proj4/
appears to fit the description on statistics canada here  https://www150.statcan.gc.ca/n1/pub/92-195-x/2011001/other-autre/mapproj-projcarte/m-c-eng.htm
}
\keyword{datasets}
