% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_correspondances_health_region_quebec.R
\docType{data}
\name{table_correspondances_health_region_quebec}
\alias{table_correspondances_health_region_quebec}
\title{table de correspondance des différents noms des health regions du québec}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 5 columns.
}
\usage{
table_correspondances_health_region_quebec
}
\description{
table de correspondance des différents noms des health regions du québec
}
\keyword{datasets}
