% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{make_plot fait les graphiques qui vont être combinés}
\usage{
make_plot(
  d,
  pop,
  left_axis_title,
  left_axis_text,
  right_axis_label,
  bottom_axis_label,
  bottom_axis_title,
  bigaxis,
  bigaxis_value = 250,
  smallaxis_value = 200,
  type = "maximum500"
)
}
\arguments{
\item{smallaxis_value}{}
}
\value{

}
\description{
make_plot fait les graphiques qui vont être combinés
}
