% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simon.R
\name{theme_simon}
\alias{theme_simon}
\title{Appliquer mon theme..}
\usage{
theme_simon(font_size = 14, font_family = "")
}
\arguments{
\item{font_size}{font size}

\item{font_family}{font type}
}
\value{

}
\description{
Appliquer mon theme..
}
\examples{
+theme_simon
}
